Move configuration specific errors in configerror
