+++ configv2/configerror/empty_test.go
@@ -12,11 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package models
+package configerror
 
-import "errors"
-
-// ErrDataTypeIsNotSupported can be returned by receiver, exporter or processor
-// factory methods that create the entity if the particular telemetry
-// data type is not supported by the receiver, exporter or processor.
-var ErrDataTypeIsNotSupported = errors.New("telemetry type is not supported")
+// Package with constants only.
